Move tests to right file
